requirements.txt
streamlit
streamlit-image-coordinates
numpy
matplotlib
Pillow
scipy
gradio
